requirements.txt
absl-py==0.12.0
astor==0.8.1
astunparse==1.6.3
attrs==20.3.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
contextvars==2.4
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
future==0.18.2
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graphviz==0.8.4
grpcio==1.36.1
h5py==2.10.0
hdf5storage==0.1.17
idna==2.10
immutables==0.15
importlib-metadata==3.7.3
importlib-resources==5.1.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mmdnn==0.3.1
mxnet==1.7.0.post2
networkx==2.5
nose==1.3.7
numpy==1.18.5
oauthlib==3.1.0
onnx==1.8.1
onnx-simplifier==0.3.3
onnx-tf==1.6.0
onnx2keras==0.0.24
onnxoptimizer==0.2.5
onnxruntime==1.7.0
opencv-python==4.5.1.48
openvino2tensorflow==1.10.6
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.1.2
promise==2.3
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytorch2keras==0.2.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
resnest==0.0.5
rsa==4.7.2
scipy==1.4.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.29.0
termcolor==1.1.0
test-generator==0.1.1
torch==1.8.0
torchvision==0.9.0
torchviz==0.0.1
tqdm==4.59.0
typeguard==2.11.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8
zipp==3.4.1
